Add
Added further UI enhancements to the home page by adding a graphic to background. Also, redesigned the "tabs" for the search sections. Reformatted form fields in the search sections. Added a footer with a email address placeholder. Email address will updated to actual working email address for reporting bugs and issues.

Associated with: #26, #70